WIP: Drop Python 3.5 support
We currently do not support Python 3.5 because of a JSON encoding
issue that was fixed in 3.6, and because Python 3 support in the
XBlock is currently somewhat moot because Open edX itself is (still) a
Python 2 platform.

Per
https://openedx.atlassian.net/wiki/spaces/OpenOPS/pages/146440579/Native+Open+edX+Ubuntu+16.04+64+bit+Installation,
Ubuntu 16.04 (Xenial Xerus) is the only supported Ubuntu platform: "It
may seem like other versions of Ubuntu will be fine, but they are
not. Only 16.04 is known to work." Supporting Python 3 on Ubuntu 16.04
would require supporting Python 3.5, because 3.5 is the only Python 3
version that is supported on Xenial.

However, whenever the next Open edX release drops, it will most likely
have to support Bionic (18.04), as by then Xenial will be (at best) 18
months from EOL.

Thus, we're unlikely to ever need to fix the encoding issue for 3.5,
so drop 3.5 from the test matrix.